requirements.txt
behave==1.2.6
behave-django==1.3.0
coverage==4.5.4
Django==2.2.6
djangorestframework==3.10.3
djangorestframework-gis==0.15
Markdown==3.1.1
parse==1.12.1
parse-type==0.5.2
pep8==1.7.1
psycopg2==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0

